requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-countries==7.6.1
django-js-asset==2.2.0
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.2
typing_extensions==4.12.2
